work in progress

+++ 2FSemesterProjekt2024/Services/EF/EFPassengerService.cs
@@ -18,11 +18,17 @@ namespace _2FSemesterProjekt2024.Services.EF
             return _context.Passengers;
         }
 
-        public IEnumerable<Passenger> GetPassengerById(int id)
+        //public IEnumerable<Passenger> GetPassengerById(int id)
+        //{
+        //    return _context.Passengers.Where(x => x.PassengerId == id);
+        //}
+
+        public Passenger GetPassengerbyId(int id)
         {
-            return _context.Passengers.Where(x => x.PassengerId == id);
+
+            return context.Passengers.Find(id);
         }
-        
+
         public void AddPassenger(Passenger passenger)
         {
             _context.Passengers.Add(passenger);
@@ -30,8 +36,11 @@ namespace _2FSemesterProjekt2024.Services.EF
         }
         public void DeletePassenger(Passenger passenger)
         {
-            _context.Passengers.Remove(passenger);
-            _context.SaveChanges();
+            if (passenger != null)
+            {
+                _context.Passengers.Remove(passenger);
+                _context.SaveChangesAsync();
+            }
         }
 
         public void UpdatePassenger(Passenger passenger)
